validation: fix inconsistent update of address index
During rebasing of the address index onto bitcoin-abc/bitcoin-abc@2f7d520, the
`pfClean` parameter was removed.  The address index code was relying on it to
update the db, which is side-effectful and not memory-only, like the rest of
the method. Running the node with `nCheckLevel == 3` would remove last
`nCheckDepth` blocks from the index.